slightly restructure the directory structure of the mobile navigation menu
