+++ frontend/src/components/Navigation/Desktop/DesktopGuestNav.tsx
